requirements.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.6
Django==3.0.8
django-celery-beat==2.0.0
django-crontab==0.7.1
django-redis==4.12.1
django-timezone-field==4.0
future==0.18.2
importlib-metadata==1.7.0
kazoo==2.7.0
kombu==4.6.11
python-crontab==2.5.1
python-dateutil==2.8.1
python-redis-lock==3.5.0
pytz==2020.1
redis==3.5.3
six==1.12.0
sortedcontainers==2.2.2
SQLAlchemy==1.3.18
sqlparse==0.3.1
vine==1.3.0
websockets==8.1
zipp==3.1.0
